recipe-api/requirements.txt
anyio==3.6.2
bson==0.5.10
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.89.1
h11==0.14.0
idna==3.4
motor==3.1.1
pydantic==1.10.4
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.7
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
